requirements.txt
pymongo
google-generativeai
PyPDF2
b2sdk
cohere
streamlit